Use recursive_diff for kubernetes
Move recursive_diff from cloudfront_distribution to
common.dict_transformations and reuse it in k8s modules
